refactor: :art: Improved the code through clippy

+++ src/syscall/process.rs
@@ -50,7 +50,7 @@ pub extern "C" fn process(request: u64, id_or_priority: u64, buf: u64, len: u64,
             // If `id_or_priority` is larger than u8::MAX, it will cause truncation.
             unsafe {
                 let data =
-                    core::slice::from_raw_parts(buf as *const u8, usize::from(len as usize));
+                    core::slice::from_raw_parts(buf as *const u8, len as usize);
                 if crate::process::create(data, id_or_priority as u8).is_err() {
                     return -1;
                 }
